detection.py
import cv2
from telegram_bot import send_alert
from datetime import datetime
import os

# Configuration
DETECTION_FOLDER = 'detections'
os.makedirs(DETECTION_FOLDER, exist_ok=True)

def detect_elephants():
    """Mock detection function - replace with your YOLO implementation"""
    cap = cv2.VideoCapture(0)  # Use webcam or RTSP stream
    
    while True:
        ret, frame = cap.read()
        if not ret:
            break

        # Replace this with actual YOLO detection
        elephant_detected = True  # Simulate detection
        
        if elephant_detected:
            timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
            img_path = f"{DETECTION_FOLDER}/elephant_{timestamp}.jpg"
            cv2.imwrite(img_path, frame)
            
            if send_alert(image_path=img_path):
                print(f"Alert sent for detection at {timestamp}")
            else:
                print("Failed to send alert")

if __name__ == "__main__":
    detect_elephants()