requirements.txt
python-telegram-bot==13.7
opencv-python>=4.5.0
python-dotenv>=0.19.0